// src/app/models/farm-campaign.ts
export interface IFarmCampaign {
    CampaignID: string;
    created_on: string;
    distribution_cost: number;
    distribution_method: string;
    event_date: string;
    event_description: string;
    event_name: string;
    farm_id: string;
    farm_name: string;
    status: string;
    user_name: string;
}